# crates/openapi.yaml
openapi: 3.1.0
info:
  title: Y-Sweet Server API
  version: 1.0.0
  description: |
    Y-Sweet is a real-time collaboration server for Yjs documents.

    This API provides endpoints for:
    - Creating and managing Yjs documents
    - Generating authentication tokens for document access
    - Real-time WebSocket synchronization
    - Asset management (images/videos) with presigned URLs
    - Document operations (copy, delete)

    **Fork Information**: This is a fork of jamsocket/y-sweet with custom extensions
    for presigned URLs, document management, and Datadog APM integration.

    ## API Audience Classification

    This API is organized by intended audience:

    ### 🔐 Admin API (Server Token Required)
    Administrative operations that should only be called from your backend server.
    These endpoints require a server-level authentication token.

    **Use cases:**
    - Creating new documents
    - Generating client authentication tokens
    - Deleting or copying documents
    - Store validation

    ### 🌐 Client API (Doc Token or Browser)
    Operations intended to be called directly from browser clients.
    These endpoints require a document-level authentication token.

    **Use cases:**
    - Reading and writing document content
    - Real-time WebSocket synchronization
    - Uploading and downloading assets

    ### 🔓 Public API (No Authentication)
    Public endpoints that don't require authentication.

    **Use cases:**
    - Health checks and monitoring
...
servers:
...
components:
  securitySchemes:
    ServerToken:
      type: http
      scheme: bearer
      description: |
        Server-level authentication token for administrative operations.
        Required for: document creation, store checks, document deletion, document copying.

    DocToken:
      type: http
      scheme: bearer
      description: |
        Document-level authentication token for document access.
        Generated via `/doc/:doc_id/auth` endpoint.
        Required for: reading/writing documents, accessing assets.

        Authorization levels:
        - `full`: Read and write access
        - `read-only`: Read-only access
  schemas:
...
    DocDeleteResponse:
      type: object
      required:
        - docId
        - dataDeleted
        - deletedAssets
        - success
      properties:
        docId:
          type: string
          description: ID of the deleted document
          example: "abc123"
        dataDeleted:
          type: boolean
          description: Whether the data.ysweet file was removed
          example: true
        deletedAssets:
          type: integer
          description: Number of asset objects removed
          example: 5
        success:
          type: boolean
          description: Overall operation success
          example: true
    DocCopyRequest:
      type: object
      required:
        - destinationDocId
      properties:
        destinationDocId:
          type: string
          description: The ID for the new copied document
          example: "new-doc-123"
    DocCopyResponse:
      type: object
      required:
        - sourceDocId
        - destinationDocId
        - success
      properties:
        sourceDocId:
          type: string
          description: ID of the source document
          example: "abc123"
        destinationDocId:
          type: string
          description: ID of the newly created document
          example: "new-doc-123"
        success:
          type: boolean
          description: Whether the copy operation succeeded
          example: true
    ContentUploadRequest:
      type: object
      required:
        - contentType
      properties:
        contentType:
          type: string
          pattern: "^(image|video)/"
          description: MIME type of the content to upload. Only image/* and video/* types are allowed.
          example: "image/png"
    ContentUploadResponse:
      type: object
      required:
        - uploadUrl
        - assetId
      properties:
        uploadUrl:
          type: string
          format: uri
          description: Presigned URL for uploading the asset
          example: "https://s3.amazonaws.com/bucket/path?signature=..."
        assetId:
          type: string
          description: Generated asset ID with file extension
          example: "clz1x2y3z4.png"
    Asset:
      type: object
      required:
        - assetId
        - downloadUrl
      properties:
        assetId:
          type: string
          description: Asset identifier (without extension)
          example: "clz1x2y3z4"
        downloadUrl:
          type: string
          format: uri
          description: Presigned URL for downloading the asset
          example: "https://s3.amazonaws.com/bucket/path?signature=..."
    AssetsResponse:
      type: object
      required:
        - assets
      properties:
        assets:
          type: array
          items:
            $ref: "#/components/schemas/Asset"
          description: List of assets associated with the document
paths:
...
  /d/{docId}:
    delete:
      operationId: deleteDocument
      summary: Delete document
      description: |
        Deletes a document and all its associated assets.
        This operation is irreversible.

        **Extension**: This is a custom endpoint (not part of upstream y-sweet).

        **Audience**: 🔐 Admin API (requires Server Token - backend only)
      tags:
        - Admin API
        - Documents
      security:
        - ServerToken: []
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: string
          description: Document identifier
          example: "abc123"
      responses:
        "200":
          description: Document deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocDeleteResponse"
        "401":
          description: Unauthorized - invalid or missing server token
        "404":
          description: Document not found
  /d/{docId}/copy:
    post:
      operationId: copyDocument
      summary: Copy document
      description: |
        Creates a copy of a document with a new document ID.
        All assets are also copied to the new document.
        The document is force-synced before copying to ensure data integrity.

        **Extension**: This is a custom endpoint (not part of upstream y-sweet).

        **Audience**: 🔐 Admin API (requires Server Token - backend only)
      tags:
        - Admin API
        - Documents
      security:
        - ServerToken: []
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: string
          description: Source document identifier
          example: "abc123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocCopyRequest"
      responses:
        "200":
          description: Document copied successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocCopyResponse"
        "401":
          description: Unauthorized - invalid or missing server token
        "404":
          description: Source document not found
        "409":
          description: Destination document ID already exists
  /d/{docId}/assets:
    post:
      operationId: generateUploadUrl
      summary: Generate presigned upload URL
      description: |
        Generates a presigned S3 URL for uploading an asset (image or video).
        The URL is valid for a limited time (typically 15 minutes).

        **Extension**: This is a custom endpoint (not part of upstream y-sweet).

        **Allowed content types**: Only `image/*` and `video/*` MIME types are permitted.

        **Audience**: 🌐 Client API (requires Doc Token - safe for browser)
      tags:
        - Client API
        - Assets
      security:
        - DocToken: []
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: string
          description: Document identifier
          example: "abc123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUploadRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentUploadResponse"
        "400":
          description: Invalid content type (not image/* or video/*)
        "401":
          description: Unauthorized - invalid or missing doc token

    get:
      operationId: listAssets
      summary: List document assets
      description: |
        Returns a list of all assets associated with the document,
        including presigned download URLs for each asset.

        **Extension**: This is a custom endpoint (not part of upstream y-sweet).

        The response includes a `Cache-Control` header with `private, max-age=30`
        to allow short-term caching of presigned URLs.

        **Audience**: 🌐 Client API (requires Doc Token - safe for browser)
      tags:
        - Client API
        - Assets
      security:
        - DocToken: []
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: string
          description: Document identifier
          example: "abc123"
      responses:
        "200":
          description: List of assets
          headers:
            Cache-Control:
              schema:
                type: string
                example: "private, max-age=30"
              description: Caching directive for presigned URLs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetsResponse"
        "401":
          description: Unauthorized - invalid or missing doc token

  # Single Document Mode Endpoints
...
  /assets:
    post:
      operationId: generateUploadUrlSingleDoc
      summary: Generate presigned upload URL (single-doc mode)
      description: |
        Generates a presigned S3 URL for uploading an asset in single-document mode.

        **Note**: This endpoint is only available when running in single-document mode.

        **Audience**: 🌐 Client API (Plane auth - safe for browser)
      tags:
        - Client API
        - Single Document Mode
      parameters:
        - name: x-verified-user-data
          in: header
          required: true
          schema:
            type: string
            example: '{"authorization": "full"}'
          description: Plane verified user data header
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUploadRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentUploadResponse"
        "400":
          description: Invalid content type
        "401":
          description: Unauthorized

    get:
      operationId: listAssetsSingleDoc
      summary: List document assets (single-doc mode)
      description: |
        Returns a list of all assets for the document in single-document mode.

        **Note**: This endpoint is only available when running in single-document mode.

        **Audience**: 🌐 Client API (Plane auth - safe for browser)
      tags:
        - Client API
        - Single Document Mode
      parameters:
        - name: x-verified-user-data
          in: header
          required: true
          schema:
            type: string
            example: '{"authorization": "full"}'
          description: Plane verified user data header
      responses:
        "200":
          description: List of assets
          headers:
            Cache-Control:
              schema:
                type: string
                example: "private, max-age=30"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetsResponse"
        "401":
          description: Unauthorized